% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wheatdb-package.R
\docType{package}
\name{wheatdb-package}
\alias{wheatdb}
\alias{wheatdb-package}
\title{wheatdb: An Amazing Shiny App}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: firstname lastname \email{your@email.com}

}
\keyword{internal}
