% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wheatdb-package.R
\docType{package}
\name{wheatdb-package}
\alias{wheatdb}
\alias{wheatdb-package}
\title{wheatdb: PKG_TITLE}
\description{
PKG_DESC.
}
\author{
\strong{Maintainer}: AUTHOR_FIRST AUTHOR_LAST \email{AUTHOR@MAIL.COM}

}
\keyword{internal}
